requirements.txt
pybids>=0.16.0
pandas>=2.0
nibabel>=5.0
joblib>=1.3
